requirements.txt
gradio
torch
torchvision
numpy
Pillow
